Move validation keywords into libcore
